SOSI: fix warnings related to 64bit
git-svn-id: https://svn.osgeo.org/gdal/trunk@35323 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/sosi/ogrsosidatasource.cpp
@@ -686,7 +686,7 @@ void OGRSOSIDataSource::buildOGRLineStringFromArc(long iSerial) {
     poLS->setNumPoints(npt);
     dth = dth / (npt-1);
 
-    long i;
+    int i;
     double dfEast = 0, dfNorth = 0;
 
     for (i=0; i<npt; i++) {
@@ -694,7 +694,7 @@ void OGRSOSIDataSource::buildOGRLineStringFromArc(long iSerial) {
         dfNorth = cN + r * sin(th1 + dth * i);
         if (dfEast != dfEast) { /* which is a wonderful property of nans */
           CPLError( CE_Warning, CPLE_AppDefined,
-                    "Calculated %lf for point %li of %i in curve %li.", dfEast, i, npt, iSerial);
+                    "Calculated %lf for point %d of %d in curve %li.", dfEast, i, npt, iSerial);
         }
         poLS->setPoint(i, dfEast, dfNorth);
     }
